requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
decorator==4.4.2
future==0.18.2
idna==2.9
lxml==4.5.0
pathspec==0.7.0
pkg-resources==0.0.0
pycparser==2.20
python-telegram-bot==12.4.2
regex==2020.2.20
requests==2.23.0
six==1.14.0
toml==0.10.0
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.8
